bounds checking is critical on kernels -- everything working well, objrec updated.

+++ examples/objrec/led_env.go
@@ -130,7 +130,7 @@ func (ev *LEDEnv) SetOutput(out int) {
 // OutErr scores the output activity of network, returning the index of
 // item with max overall activity, and 1 if that is error, 0 if correct.
 // also returns a top-two error: if 2nd most active output was correct.
-func (ev *LEDEnv) OutErr(tsr *tensor.Float32, corLED int) (maxi int, err, err2 float64) {
+func (ev *LEDEnv) OutErr(tsr *tensor.Float64, corLED int) (maxi int, err, err2 float64) {
 	nc := ev.Output.Len() / ev.NOutPer
 	maxi = 0
 	maxv := 0.0
